// src/modules/smart-contract-case/dto/update-case-nft.dto.ts
import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";

export class UpdateCaseNftDto {
    @ApiProperty({ type: String, description: 'This is a required property' })
    title: string = "";

    @ApiProperty({ type: String, description: 'This is a required property' })
    description: string = "";

    @ApiProperty({ type: String, description: 'This is a required property' })
    category: string = "";

    @ApiProperty({ type: String, description: 'This is a required property' })
    owner: string = "";

    @ApiProperty({ type: Number, description: 'This is a required property' })
    bounty: number = 0;

    @ApiProperty({ type: String, description: 'This is a required property' })
    file: string = "";

    @ApiProperty({ type: String, description: 'This is a required property' })
    status: string = "";
}